Getting product images with Picasso

+++ app/src/main/java/com/example/e_commerceapp/Models/ProductModel.kt
@@ -1,9 +1,11 @@
 package com.example.e_commerceapp.Models
 
+import android.net.Uri
+
 data class ProductModel (
     val id:Int,
     val title:String,
-    val price:Int,
+    val price:Float,
     val description:String,
     val image:String,
